pyproject.toml
[tool.poetry]
name = "incident api"
version = "0.1.0"
description = "Test task for UCAR-TOPDOER"
authors = ["Tychin Denis"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
django = ">=5.2,<6.0"
d = "^0.2.2"
djangorestframework = ">=3.16.0,<4.0.0"
drf-spectacular = "^0.27.1"
djoser = ">=2.3.1,<3.0.0"
django-cors-headers = ">=4.7.0,<5.0.0"
drf-standardized-errors = ">=0.14.1,<0.15.0"
django-health-check = "^3.19.1"
django-debug-toolbar = ">=5.2.0,<6.0.0"
nplusone = ">=1.0.0,<2.0.0"
django-telegram-logging = ">=1.0.2,<2.0.0"
django-structlog = ">=9.1.1,<10.0.0"
python-decouple = "^3.8"
psycopg = "^3.2.7"
gunicorn = "23.0.0"
django-querycount = "^0.8.3"
django-query-counter = "^0.4.1"
whitenoise = "^6.9.0"
safety = "^3.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"